refactor: convert codebase to TypeScript
BREAKING CHANGE: Project now requires TypeScript compilation step
- Added type definitions for all components
- Updated build process to include TypeScript compilation
- Added tsconfig.json and necessary type dependencies
