--- src/config/schema.json
@@ -1,3 +1,0 @@
-{
-  "type": "string"
-}
