+++ src/core/LicenseService.ts
